+++ solr/core/src/test/org/apache/solr/util/cli/AuthToolTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.solr.util;
+package org.apache.solr.util.cli;
 
 import static org.apache.solr.util.SolrCLI.findTool;
 import static org.apache.solr.util.SolrCLI.parseCmdLine;
@@ -25,6 +25,7 @@ import java.nio.file.Path;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.io.file.PathUtils;
 import org.apache.solr.cloud.SolrCloudTestCase;
+import org.apache.solr.util.SolrCLI;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -80,7 +81,7 @@ public class AuthToolTest extends SolrCloudTestCase {
   }
 
   private int runTool(String[] args) throws Exception {
-    SolrCLI.Tool tool = findTool(args);
+    Tool tool = findTool(args);
     assertTrue(tool instanceof SolrCLI.AuthTool);
     CommandLine cli = parseCmdLine(tool.getName(), args, tool.getOptions());
     return tool.runTool(cli);
